Create GreatCircleDistance.java
The great-circle distance is the length of the shortest path between two points (x1,y1) and (x2,y2) on the surface of a sphere. 
Write a Java program that takes four double input numbers x1, y1, x2, y2, which are the latitude and longitude in degrees of two points on the surface of the Earth;
uses the mean radius of the Earth r = 6,371.0  kilometres;
and prints the great-circle distance dist between them in kilometres.
Note that the input numbers are given in degrees but Java’s trigonometric functions use radians. Use Math.toRadians() to convert from degrees to radians.